+++ src/main/scala/gr/grnet/aquarium/logic/accounting/dsl/DSLPriceList.scala
@@ -43,7 +43,7 @@ package gr.grnet.aquarium.logic.accounting.dsl
 case class DSLPriceList (
   override val name: String,
   override val overrides: Option[DSLPriceList],
-  prices: Map[DSLResource,  Float],
+  prices: Map[DSLResource,  Double],
   override val effective: DSLTimeFrame
 ) extends DSLTimeBoundedItem[DSLPriceList](name, overrides, effective) {
